+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to the "digital-asm" extension will be documented in this fi
 
 * added extension entry point 
 * added properties for accessing the digital simulator and asm3 programm
+* added basis for test suite
 
 ## 0.1.0 - 01.12.2019